requirements.txt
alembic==0.9.6
aniso8601==1.3.0
attrs==17.3.0
Babel==2.5.1
blinker==1.4
certifi==2017.11.5
chardet==3.0.4
ciso8601==1.0.5
click==6.7
contextlib2==0.5.5
coverage==4.4.2
croniter==0.3.20
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-RQ2==17.2
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-SSE==0.2.1
Flask-WTF==0.14.2
funcsigs==1.0.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
hashids==1.2.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
msgpack-python==0.4.8
olefile==0.44
passlib==1.7.1
Pillow==4.3.0
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
PyJWT==1.5.3
pytest==3.3.1
pytest-cov==2.5.1
pytest-cover==3.0.0
pytest-coverage==0.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
raven==6.3.0
redis==2.10.6
requests==2.18.4
rq==0.9.2
rq-scheduler==0.7.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.1.15
sqlalchemy-postgres-copy==0.5.0
unicodecsv==0.14.1
urllib3==1.22
utm==0.4.2
Werkzeug==0.13
WTForms==2.1
